+++ app/src/main/java/br/digitalhouse/marveltime/view/adapter/AdapterRecyclerPersonagens.java
@@ -1,18 +1,12 @@
 package br.digitalhouse.marveltime.view.adapter;
-import android.content.Context;
-import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
-import java.util.ArrayList;
 import java.util.List;
-
 import br.digitalhouse.marveltime.model.PersonagemResult;
 import br.digitalhouse.marveltime.view.Interfaces.OnClick;
-import br.digitalhouse.marveltime.view.activity.PersonagensTelaActivity;
-import br.digitalhouse.marveltime.model.CardModel;
 import br.digitalhouse.marveltime.R;
 import br.digitalhouse.marveltime.view.viewholder.ViewHolderPersonagens;
